pyproject.toml
[tool.poetry]
name = "chat-langchain"
version = "0.1.0"
description = ""
authors = ["CropGuard"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = ">=1.9.0,<2.0.0"
fastapi = "^0.103.1"
pydantic = ">=2.5.0,<3.0.0"
langchain = ">=0.1.1,<0.2.0"
uvicorn = "^0.23.2"
beautifulsoup4 = "^4.12.2"
tiktoken = ">=0.5.2,<0.6.0"
weaviate-client = "^4.4.4"
psycopg2 = "^2.9.7"
lxml = "^4.9.3"
langserve = {extras = ["server"], version = "^0.0.34"}
voyageai = "^0.1.4"
PyPDF2 = "3.0.1"
boto3 = "1.34.69"
python-dotenv = "1.0.0"
chromadb = ">=0.4.22,<0.5"
sentence-transformers = "2.2.2"
lark = "1.1.8"
langsmith = ">=0.0.83,<0.1"
langchain-core = ">=0.1.13,<0.2"
# dsmain = "0.0.1"
langchain-openai = ">=0.0.3,<0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

